+++ js/util_getId.js
@@ -1,0 +1,20 @@
+import {getRandomInteger} from './util.js';
+// получаем случайное неповторяющееся число.
+const getId = (min, max) => {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+
+
+export {getId};
